BE/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-taggit==5.0.1
django-filter==23.5
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
